requirements.txt
black==24.2.0
connexion[swagger-ui]==2.14.1
flake8==7.0.0
Flask==2.2.2
greenlet==2.0.2
mysql-connector-python==8.3.0
numpy==1.25.2
openpyxl==3.1.2
pandas==1.5.0
pendulum==2.1.2
psycopg2==2.9.7
pymongo==4.6.3
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
vulture==2.11